feat(reorg): reorganize names
