--- aliquotmaf/subcommands/merge_protected/runners/__init__.py
@@ -1,9 +1,0 @@
-from __future__ import absolute_import
-
-from .base import BaseRunner
-from .gdc_1_0_0_merged_protected import GDC_1_0_0_Merged_Protected
-
-__all__ = [
-    BaseRunner,
-    GDC_1_0_0_Merged_Protected
-]
