--- aliquotmaf/subcommands/protected_to_public/runners/__init__.py
@@ -1,9 +1,0 @@
-from __future__ import absolute_import
-
-from .base import BaseRunner
-from .gdc_1_0_0_merged_public import GDC_1_0_0_Merged_Public
-
-__all__ = [
-    BaseRunner,
-    GDC_1_0_0_Merged_Public
-]
