--- aliquotmaf/subcommands/vcf_to_protected/extractors/__init__.py
@@ -1,18 +1,0 @@
-from __future__ import absolute_import
-
-from .base import Extractor
-from .genotypes import VariantAlleleIndexExtractor, GenotypeAndDepthsExtractor
-from .location import LocationDataExtractor
-from .effects import EffectsExtractor, SelectOneEffectExtractor
-from .population_frequency import PopulationFrequencyExtractor
-from .variant_class import VariantClassExtractor
-
-__all__ = [
-    Extractor,
-    VariantAlleleIndexExtractor,
-    GenotypeAndDepthsExtractor,
-    LocationDataExtractor,
-    EffectsExtractor,
-    SelectOneEffectExtractor,
-    PopulationFrequencyExtractor,
-    VariantClassExtractor]
