--- aliquotmaf/subcommands/vcf_to_protected/runners/__init__.py
@@ -1,9 +1,0 @@
-from __future__ import absolute_import
-
-from .base import BaseRunner
-from .gdc_1_2_0_protected import GDC_1_2_0_Protected
-
-__all__ = [
-    BaseRunner,
-    GDC_1_2_0_Protected
-]
