+++ requirements.txt
@@ -1,2 +1,2 @@
 pysam==0.14.1
-git+ssh://git@github.com/NCI-GDC/maf-lib.git@7693524eb0951a13f2e48f3bd1f43578b0c8d40f#egg=maf-lib
+git+ssh://git@github.com/NCI-GDC/maf-lib.git@8cf0a7cf4252289caeca406b3df9f8cfb2a89413#egg=maf-lib
