+++ tests/extractors/test_effects.py
@@ -3,7 +3,7 @@ Tests the extractors in aliquotmaf.subcommands.vcf_to_protected.extractors.effec
 """
 import pytest
 
-from aliquotmaf.subcommands.vcf_to_protected.extractors.effects import EffectsExtractor, SelectOneEffectExtractor
+from aliquotmaf.subcommands.vcf_to_aliquot.extractors.effects import EffectsExtractor, SelectOneEffectExtractor
 
 # VariantAlleleIndexExtractor -> GenotypeAndDepthsExtractor -> LocationDataExtractor -> EffectsExtractor -> SelectOneEffectExtractor -> PopulationFrequencyExtractor -> VariantClassExtractor
